+++ components/pages/MypageEdit/Introduction/styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { theme } from "../../../styles/Theme";
+import { theme } from "../../../../styles/Theme";
 
 export const ProfileWrapper = styled.div`
   width: 100%;
