+++ components/pages/MypageEdit/Skill/styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { theme } from "../../../styles/Theme";
+import { theme } from "../../../../styles/Theme";
 
 export const TagWrapper = styled.div`
   width: 100%;
